# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Monark API
  version: 0.1.0
  description: Developer hub API for project monitoring, dependency tracking, and technical assessments.
servers:
  - url: http://localhost:8000/api
    description: Development
  - url: https://api.monark.dev/api
    description: Production
paths:
...
  /dependency/dependencies:
    get:
      summary: List dependencies
      operationId: listDependencies
      tags:
        - Dependency
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Paginated dependencies list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedDependenciesResponse"

    post:
      summary: Create a dependency
      operationId: createDependency
      tags:
        - Dependency
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDependencyInput"
      responses:
        "201":
          description: Dependency created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DependencyResponse"
...
components:
  schemas:
    DependencyOutput:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        currentVersion:
          type: string
        latestVersion:
          type: string
        ltsVersion:
          type: string
        packageManager:
          type: string
          enum: [composer, npm, pip]
        type:
          type: string
          enum: [runtime, dev]
        isOutdated:
          type: boolean
        projectId:
          type: string
          format: uuid
        repositoryUrl:
          type: string
          format: uri
          nullable: true
          maxLength: 2048
        vulnerabilityCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DependencyResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/DependencyOutput"
    PaginatedDependenciesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/DependencyOutput"
            total:
              type: integer
            page:
              type: integer
            per_page:
              type: integer
            total_pages:
              type: integer
    CreateDependencyInput:
      type: object
      required:
        - name
        - currentVersion
        - latestVersion
        - ltsVersion
        - packageManager
        - type
        - isOutdated
        - projectId
      properties:
        name:
          type: string
          maxLength: 255
        currentVersion:
          type: string
          maxLength: 100
        latestVersion:
          type: string
          maxLength: 100
        ltsVersion:
          type: string
          maxLength: 100
        packageManager:
          type: string
          enum: [composer, npm, pip]
        type:
          type: string
          enum: [runtime, dev]
        isOutdated:
          type: boolean
        projectId:
          type: string
          format: uuid
        repositoryUrl:
          type: string
          format: uri
          nullable: true
          maxLength: 2048
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HMAC-SHA256
security:
  - bearerAuth: []
...